Added search logic in maids list, added the cases in async task for server connection and updated interface with APIs. Added progress dialog for server connection waiting for every request.

+++ app/src/main/java/com/hack/xapp/model/Coordinate.java
@@ -5,10 +5,10 @@ package com.hack.xapp.model;
  */
 public class Coordinate {
 
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public Coordinate(int cx, int cy) {
+    public Coordinate(double cx, double cy) {
         x = cx;
         y = cy;
     }
